requirements.txt
aerich==0.7.2
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiomysql==0.2.0
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dictdiffer==0.9.0
fastapi==0.115.4
frozenlist==1.5.0
h11==0.14.0
httptools==0.6.4
idna==3.10
iso8601==2.1.0
multidict==6.1.0
nanoid==2.0.0
openai==0.28.0
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
PyMySQL==1.1.1
pypika-tortoise==0.2.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
requests==2.32.3
sniffio==1.3.1
starlette==0.41.2
tomlkit==0.13.2
tortoise-orm==0.21.7
tqdm==4.66.6
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
yarl==1.17.1
